Prevent multiple WebEventListeners from being added to WebDriverRunner
When running the PuppyAdoptionTests through a debugger, I was able to
discover that as each test method was run, the amount of attached
listeners on the WebDriverRunner increased by one. As this happened,
logs WebEventListener logs would increase according to the amount of
listeners that were registered.

This commit changes it so that the listener only gets added once per
class.